# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Deve Sub API",
    "description": "Self-hosted proxy subscription infrastructure manager",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/audit-logs": {
      "get": {
        "summary": "`GET /api/v1/audit-logs` — list audit log entries with filters and\ncursor pagination (admin only). AUDIT-001.",
        "operationId": "list_audit_logs",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Max entries per page (1-100, default 50)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Pagination cursor (oldest entry ULID from previous page)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actor_id",
            "in": "query",
            "description": "Filter by actor ULID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "in": "query",
            "description": "Filter by action (e.g. \"auth.login\")",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "target_type",
            "in": "query",
            "description": "Filter by target type (e.g. \"user\")",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "target_id",
            "in": "query",
            "description": "Filter by target ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "since",
            "in": "query",
            "description": "Inclusive RFC3339 timestamp (whole seconds)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "before",
            "in": "query",
            "description": "Exclusive RFC3339 timestamp (whole seconds)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audit log entries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListAuditLogsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid cursor or actor_id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie_auth": []
          }
        ]
      }
    },
    "/api/v1/audit-logs/cleanup": {
      "post": {
        "operationId": "cleanup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditCleanupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditCleanupResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie_auth": []
          }
        ]
      }
    },
    "/api/v1/audit-logs/cleanup/preview": {
      "post": {
        "operationId": "preview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditCleanupPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditCleanupPreviewResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie_auth": []
          }
        ]
      }
    },
    "/api/v1/audit-logs/policy": {
      "get": {
        "operationId": "policy",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditPolicyResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditCleanupPreviewRequest": {
        "type": "object",
        "description": "Select a cleanup cutoff using the server clock.",
        "required": [
          "keep_days"
        ],
        "properties": {
          "keep_days": {
            "type": "integer",
            "format": "int32",
            "description": "Whole days to keep, between 1 and 3650.",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
      "AuditCleanupPreviewResponse": {
        "type": "object",
        "description": "Exact bounded snapshot that must be confirmed before deletion.",
        "required": [
          "before_unix_ms",
          "entry_ids",
          "has_more"
        ],
        "properties": {
          "before_unix_ms": {
            "type": "integer",
            "format": "int64"
          },
          "entry_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "has_more": {
            "type": "boolean"
          }
        }
      },
      "AuditCleanupRequest": {
        "type": "object",
        "description": "Confirm exactly the IDs and cutoff returned by preview.",
        "required": [
          "before_unix_ms",
          "entry_ids"
        ],
        "properties": {
          "before_unix_ms": {
            "type": "integer",
            "format": "int64"
          },
          "entry_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
      "AuditCleanupResponse": {
        "type": "object",
        "description": "Committed cleanup result; the receipt is queryable in the audit log.",
        "required": [
          "deleted",
          "receipt_id"
        ],
        "properties": {
          "deleted": {
            "type": "integer",
            "minimum": 0
          },
          "receipt_id": {
            "type": "string"
          }
        }
      },
      "AuditLogDto": {
        "type": "object",
        "description": "A single audit log entry.",
        "required": [
          "id",
          "action",
          "created_at"
        ],
        "properties": {
          "action": {
            "type": "string",
            "description": "Action string (e.g. `\"auth.login\"`, `\"user.create\"`)."
          },
          "actor_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "The actor's ULID, or `null` for system/anonymous actions."
          },
          "created_at": {
            "type": "string",
            "description": "When the action was recorded (ISO 8601 UTC)."
          },
          "details_json": {
            "type": [
              "string",
              "null"
            ],
            "description": "Non-sensitive metadata as a JSON string."
          },
          "id": {
            "type": "string",
            "description": "ULID identifier."
          },
          "target_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Target entity ULID."
          },
          "target_type": {
            "type": [
              "string",
              "null"
            ],
            "description": "Target entity type (e.g. `\"user\"`, `\"source\"`)."
          }
        }
      },
      "AuditPolicyResponse": {
        "type": "object",
        "description": "Effective server-owned retention policy; zero disables automatic expiry.",
        "required": [
          "retention_days",
          "batch_limit"
        ],
        "properties": {
          "batch_limit": {
            "type": "integer",
            "minimum": 0
          },
          "retention_days": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response for auth endpoints.",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Machine-readable error code (e.g. `\"unauthorized\"`, `\"forbidden\"`)."
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message."
          }
        }
      },
...
      "ListAuditLogsResponse": {
        "type": "object",
        "description": "Response body for `GET /api/v1/audit-logs` (cursor-paginated audit log).",
        "required": [
          "entries"
        ],
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditLogDto"
            },
            "description": "Audit log entries in the current page (newest first)."
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Cursor for the next page (`None` if no more results). The cursor is\nthe oldest entry's ULID in the current page."
          }
        }
      },
...
    },
    "securitySchemes": {
      "cookie_auth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "deve_sub_session"
      }
    }
  }
}